+++ docs/data/charts/bars/BarDownsamplingBasic.js
@@ -5,29 +5,28 @@ import { BarChart } from '@mui/x-charts/BarChart';
 const generateLargeDataset = (size) => {
   return Array.from({ length: size }, (_, i) => ({
     x: i,
-    value: Math.sin(i / 50) * 100 + 50 + 200,
+    value: Math.sin(i / 50) * 100 + 200,
   }));
 };
 
-const largeDataset = generateLargeDataset(1000);
+const largeDataset = generateLargeDataset(500);
 const xAxisData = largeDataset.map((item) => item.x);
 const yAxisData = largeDataset.map((item) => item.value);
 
 export default function BarDownsamplingBasic() {
   return (
-    <div style={{ width: '100%', height: '400px' }}>
-      <h3>Original Data (1000 points)</h3>
+    <div>
+      <h3>Original Data (500 points)</h3>
       <BarChart
         width={600}
         height={180}
         series={[
           {
             data: yAxisData,
-            label: 'All 5000 points',
+            label: 'All 500 points',
           },
         ]}
         xAxis={[{ data: xAxisData, scaleType: 'band' }]}
-        margin={{ left: 60, right: 20, top: 20, bottom: 40 }}
       />
       <h3>Downsampled Data (Auto - 200 points)</h3>
       <BarChart
@@ -41,7 +40,6 @@ export default function BarDownsamplingBasic() {
         ]}
         xAxis={[{ data: xAxisData, scaleType: 'band' }]}
         downsample={true} // Enable automatic downsampling at chart level
-        margin={{ left: 60, right: 20, top: 20, bottom: 40 }}
       />
     </div>
   );
