+++ docs/data/charts/bars/BarDownsamplingCustom.js
@@ -6,11 +6,11 @@ const generateTimeSeriesData = (size) => {
   const baseTime = new Date('2023-01-01').getTime();
   return Array.from({ length: size }, (_, i) => ({
     timestamp: new Date(baseTime + i * 86400000), // Daily data
-    value: Math.sin(i / 30) * 50 + Math.random() * 30 + 100,
+    value: Math.sin(i / 30) * 50 * 30 + 100,
   }));
 };
 
-const timeSeriesData = generateTimeSeriesData(2000);
+const timeSeriesData = generateTimeSeriesData(1000);
 
 // Custom downsampling function that prioritizes recent data
 const customDownsample = (data, targetPoints) => {
@@ -34,7 +34,7 @@ const customDownsample = (data, targetPoints) => {
 };
 
 // Another custom function: keep every Nth point with some randomness
-const randomizedDownsample = (data, targetPoints) => {
+const everyNth = (data, targetPoints) => {
   if (data.length <= targetPoints) {
     return [...data];
   }
@@ -44,7 +44,7 @@ const randomizedDownsample = (data, targetPoints) => {
 
   for (let i = 0; i < targetPoints; i += 1) {
     const baseIndex = Math.floor(i * step);
-    const randomOffset = Math.floor(Math.random() * step * 0.3);
+    const randomOffset = Math.floor(step * 0.3);
     const index = Math.min(baseIndex + randomOffset, data.length - 1);
     result.push(data[index]);
   }
@@ -58,7 +58,7 @@ const xAxisData = timeSeriesData.map((_, i) => i);
 export default function BarDownsamplingCustom() {
   return (
     <div style={{ width: '100%' }}>
-      <h3>Original Data (2000 points)</h3>
+      <h3>Original Data (1000 points)</h3>
       <BarChart
         width={800}
         height={200}
@@ -96,7 +96,7 @@ export default function BarDownsamplingCustom() {
           },
         ]}
         xAxis={[{ data: xAxisData, scaleType: 'band' }]}
-        downsample={randomizedDownsample}
+        downsample={everyNth}
         margin={{ left: 60, right: 20, top: 20, bottom: 40 }}
       />
     </div>
