+++ docs/data/charts/bars/BarDownsamplingCustom.tsx
@@ -34,7 +34,7 @@ const customDownsample = (data, targetPoints) => {
 };
 
 // Another custom function: keep every Nth point with some randomness
-const randomizedDownsample = (data, targetPoints) => {
+const everyNth = (data, targetPoints) => {
   if (data.length <= targetPoints) {
     return [...data];
   }
@@ -44,7 +44,7 @@ const randomizedDownsample = (data, targetPoints) => {
 
   for (let i = 0; i < targetPoints; i += 1) {
     const baseIndex = Math.floor(i * step);
-    const randomOffset = Math.floor(Math.random() * step * 0.3);
+    const randomOffset = Math.floor(step * 0.3);
     const index = Math.min(baseIndex + randomOffset, data.length - 1);
     result.push(data[index]);
   }
@@ -98,7 +98,7 @@ export default function BarDownsamplingCustom() {
           },
         ]}
         xAxis={[{ data: xAxisData, scaleType: 'band' }]}
-        downsample={randomizedDownsample}
+        downsample={everyNth}
         margin={{ left: 60, right: 20, top: 20, bottom: 40 }}
       />
     </div>
