+++ docs/data/charts/bars/BarDownsamplingStrategies.js
@@ -4,15 +4,15 @@ import { BarChart } from '@mui/x-charts/BarChart';
 // Generate dataset with varying patterns to showcase different strategies
 const generateDataWithPattern = (size) => {
   return Array.from({ length: size }, (_, i) => {
-    const baseValue = Math.sin(i / 100) * 50 + 100;
-    const spike = i % 200 === 0 ? 150 : 0; // Periodic spikes
-    const noise = Math.random() * 20 - 10;
+    const baseValue = Math.sin(i / 100) * 50 + 60;
+    const spike = i % 100 === 0 ? 50 : 0;
+    const noise = i % 2 === 0 ? 10 : -10;
     return baseValue + spike + noise;
   });
 };
 
-const data = generateDataWithPattern(3000);
-const xAxisData = Array.from({ length: 3000 }, (_, i) => i);
+const data = generateDataWithPattern(500);
+const xAxisData = Array.from({ length: 500 }, (_, i) => i);
 
 export default function BarDownsamplingStrategies() {
   return (
@@ -28,8 +28,7 @@ export default function BarDownsamplingStrategies() {
           },
         ]}
         xAxis={[{ data: xAxisData, scaleType: 'band' }]}
-        downsample={{ targetPoints: 300, strategy: 'linear' }}
-        margin={{ left: 60, right: 20, top: 20, bottom: 40 }}
+        downsample={{ targetPoints: 50, strategy: 'linear' }}
       />
       <h3>Max Strategy (Preserves Peaks)</h3>
       <BarChart
@@ -42,8 +41,7 @@ export default function BarDownsamplingStrategies() {
           },
         ]}
         xAxis={[{ data: xAxisData, scaleType: 'band' }]}
-        downsample={{ targetPoints: 300, strategy: 'max' }}
-        margin={{ left: 60, right: 20, top: 20, bottom: 40 }}
+        downsample={{ targetPoints: 50, strategy: 'max' }}
       />
       <h3>Average Strategy (Smooths Data)</h3>
       <BarChart
@@ -56,8 +54,7 @@ export default function BarDownsamplingStrategies() {
           },
         ]}
         xAxis={[{ data: xAxisData, scaleType: 'band' }]}
-        downsample={{ targetPoints: 300, strategy: 'average' }}
-        margin={{ left: 60, right: 20, top: 20, bottom: 40 }}
+        downsample={{ targetPoints: 50, strategy: 'average' }}
       />
       <h3>Peak Strategy (Preserves Important Variations)</h3>
       <BarChart
@@ -70,8 +67,7 @@ export default function BarDownsamplingStrategies() {
           },
         ]}
         xAxis={[{ data: xAxisData, scaleType: 'band' }]}
-        downsample={{ targetPoints: 300, strategy: 'peak' }}
-        margin={{ left: 60, right: 20, top: 20, bottom: 40 }}
+        downsample={{ targetPoints: 50, strategy: 'peak' }}
       />
     </div>
   );
